move utility functions to the utils folder update file path for the util functions tests add terraform to create a package for utils add utils package to ingester lambda layer
