+++ src/ingester/utils/requirements.txt
@@ -1,0 +1,8 @@
+numpy==1.26.1
+pandas==2.1.2
+pg8000==1.30.2
+python-dateutil==2.8.2
+pytz==2023.3.post1
+scramp==1.4.4
+six==1.16.0
+tzdata==2023.3
