+++ test/test_ingester/test_dump_data.py
@@ -2,7 +2,7 @@ import os
 import pytest
 import boto3
 from moto import mock_s3
-from src.ingester.dump_data import dump_data
+from ingester.utils.dump_data import dump_data
 
 
 @pytest.fixture(scope="function")
